+++ packages/@markuplint/rules/src/create-message.ts
@@ -253,7 +253,7 @@ function createExpectedObject(
 			? null
 			: 1 < expectedObject.length
 				? t('either {0}', t(expectedObject))
-				: expectedObject[0] ?? null;
+				: (expectedObject[0] ?? null);
 
 	return expects;
 }