+++ packages/@markuplint/svelte-parser/src/parser.ts
@@ -113,7 +113,7 @@ export class SvelteParser extends Parser<SvelteNode> {
 				const reEndTag = new RegExp(`</${originNode.name}\\s*>$`, 'i');
 				const startTagEndOffset =
 					children.length > 0
-						? children[0]?.start ?? 0
+						? (children[0]?.start ?? 0)
 						: token.raw.replace(reEndTag, '').length + token.startOffset;
 				const startTagLocation = this.sliceFragment(token.startOffset, startTagEndOffset);