+++ packages/@markuplint/i18n/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.1](https://github.com/markuplint/markuplint/compare/@markuplint/i18n@4.5.0...@markuplint/i18n@4.5.1) (2024-06-25)
+
+**Note:** Version bump only for package @markuplint/i18n
+
 # [4.5.0](https://github.com/markuplint/markuplint/compare/@markuplint/i18n@4.5.0-alpha.0...@markuplint/i18n@4.5.0) (2024-05-04)
 
 **Note:** Version bump only for package @markuplint/i18n