+++ packages/@markuplint/pretenders/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.7](https://github.com/markuplint/markuplint/compare/@markuplint/pretenders@0.0.6...@markuplint/pretenders@0.0.7) (2024-06-25)
+
+**Note:** Version bump only for package @markuplint/pretenders
+
+## [0.0.6](https://github.com/markuplint/markuplint/compare/@markuplint/pretenders@0.0.5...@markuplint/pretenders@0.0.6) (2024-06-09)
+
+**Note:** Version bump only for package @markuplint/pretenders
+
 ## [0.0.5](https://github.com/markuplint/markuplint/compare/@markuplint/pretenders@0.0.4...@markuplint/pretenders@0.0.5) (2024-05-28)
 
 **Note:** Version bump only for package @markuplint/pretenders