+++ packages/@markuplint/shared/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.3](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.2...@markuplint/shared@4.4.3) (2024-06-25)
+
+**Note:** Version bump only for package @markuplint/shared
+
+## [4.4.2](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.1...@markuplint/shared@4.4.2) (2024-06-09)
+
+### Bug Fixes
+
+- fix to export type files ([eff4bbf](https://github.com/markuplint/markuplint/commit/eff4bbfd127574809dc5e15d7cafe87699758ee0))
+
 ## [4.4.1](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.0...@markuplint/shared@4.4.1) (2024-05-28)
 
 **Note:** Version bump only for package @markuplint/shared