+++ packages/markuplint/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.2](https://github.com/markuplint/markuplint/compare/markuplint@4.9.1...markuplint@4.9.2) (2024-06-25)
+
+### Bug Fixes
+
+- **markuplint:** ensure `--config` option correctly handles absolute paths ([fd36c1a](https://github.com/markuplint/markuplint/commit/fd36c1a4475f71d8b1e83212a60639c7361cd702))
+
+## [4.9.1](https://github.com/markuplint/markuplint/compare/markuplint@4.9.0...markuplint@4.9.1) (2024-06-09)
+
+### Bug Fixes
+
+- fix to export type files ([eff4bbf](https://github.com/markuplint/markuplint/commit/eff4bbfd127574809dc5e15d7cafe87699758ee0))
+
 # [4.9.0](https://github.com/markuplint/markuplint/compare/markuplint@4.8.1...markuplint@4.9.0) (2024-05-28)
 
 ### Features